Move types to types.d.ts and reorganize components
