+++ components/UI/ClearCartButton.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { handleClearCart } from "../../app/[locale]/actions";
+import { handleClearCart } from "../../app/actions";
 import { useI18n } from "../../locales/client";
 
 export default function ClearCartButton() {
