--- scripts/login.ts
@@ -1,10 +1,0 @@
-// export const handleLogin = async (username: string, password: string) => {
-//   await fetch(`${process.env.NEXT_PUBLIC_VERCEL_URL}/api/login`, {
-//     method: "POST",
-//     headers: { "Content-Type": "application/json" },
-//     body: JSON.stringify({
-//       username,
-//       password,
-//     }),
-//   });
-// };
